requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.5
django-crispy-forms==1.14.0
django-environ==0.9.0
django-heroku==0.3.1
django-nocaptcha-recaptcha==0.0.20
django-recaptcha==3.0.0
gunicorn==20.1.0
psycopg2==2.9.3
python-decouple==3.6
python-dotenv==0.20.0
smartencoding==0.3
sqlparse==0.4.2
tzdata==2021.5
whitenoise==6.0.0
